PaddedLeft, mode indicator, pitch fix

+++ lib/Telemetry/src/Telemetry.h
@@ -15,7 +15,7 @@ class Telemetry {
         bool antenna;
 
         float twr;
-        byte pitch;
+        int pitch;
 
         long altitude;
         long periapsis;
